src/requirements.txt
python-dotenv
aiohttp
asyncio
https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
tgcrypto